refactored the project and tried to make clean-code

+++ src/main/java/com/example/asyncnotemanagerapi/dto/NoteResponseDTO.java
@@ -4,7 +4,7 @@ import java.time.LocalDateTime;
 import java.util.List;
 
 public record NoteResponseDTO(
-        int id,
+        Long id,   // ✅ Long olmalı
         String title,
         String content,
         String category,
