+++ src/main/java/com/example/asyncnotemanagerapi/repository/NoteRepository.java
@@ -7,9 +7,8 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface NoteRepository extends JpaRepository<Note, Integer> {
+public interface NoteRepository extends JpaRepository<Note, Long> {
 
     // Kategoriye göre filtreleme
     List<Note> findByCategory(String category);
-
 }
