+++ src/main/java/com/example/asyncnotemanagerapi/service/NoteService.java
@@ -9,30 +9,21 @@ import java.util.concurrent.CompletableFuture;
 
 public interface NoteService {
 
-    // Tüm notları getir (veritabanından)
     List<NoteResponseDTO> getAllNotes();
 
-    // ID ile tek bir not getir
-    NoteResponseDTO getNoteById(int id);
+    NoteResponseDTO getNoteById(Long id);
 
-    // Yeni not oluştur ve kaydet
     NoteResponseDTO createNote(NoteRequestDTO request);
 
-    // Var olan notu güncelle
-    NoteResponseDTO updateNote(int id, NoteRequestDTO request);
+    NoteResponseDTO updateNote(Long id, NoteRequestDTO request);
 
-    // Notu sil
-    void deleteNote(int id);
+    void deleteNote(Long id);
 
-    // Kategoriye göre filtreleme
     List<NoteResponseDTO> filterByCategory(String category);
 
-    // Başlığa veya tarihe göre sıralama
     List<NoteResponseDTO> sortNotes(String criteria);
 
-    // Asenkron yedekleme başlat
     CompletableFuture<Void> backupNotesAsync();
 
-    // Son yedekleme durumunu döner
     BackupStatus getBackupStatus();
 }
